quarkus: consolidate kamelet integration tests
