+++ .github/workflows/ci-build.yml
@@ -111,7 +111,6 @@ jobs:
         native-image-project:
           - :camel-k-quarkus-itests-core
           - :camel-k-quarkus-itests-runtime
-          - :camel-k-quarkus-itests-runtime-kamelet
           - :camel-k-quarkus-itests-cron
           - :camel-k-quarkus-itests-master
           - :camel-k-quarkus-itests-kamelet
