+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-kamelet/src/main/resources/application.properties
@@ -18,10 +18,6 @@
 #
 # Quarkus
 #
+quarkus.log.console.enable = false
 quarkus.banner.enabled     = false
-
-#
-# Quarkus :: Camel
-#
-quarkus.camel.routes-discovery.enabled = false
