+++ camel-k-quarkus/camel-k-quarkus-itests/camel-k-quarkus-itests-kamelet/src/test/resources/routes.properties
@@ -22,9 +22,12 @@
 camel.k.sources[0].location          = file:{{env:ROUTES_DIR}}/set-body.yaml
 camel.k.sources[0].type              = template
 camel.k.sources[0].property-names[0] = bodyValue
+camel.k.sources[1].location          = file:{{env:ROUTES_DIR}}/to-upper.yaml
+camel.k.sources[1].type              = template
+camel.k.sources[1].property-names[0] = message
 
 #
 # camel-k - sources (routes)
 #
-camel.k.sources[1].location = file:{{env:ROUTES_DIR}}/route.yaml
-camel.k.sources[1].type     = source
+camel.k.sources[2].location          = file:{{env:ROUTES_DIR}}/process.yaml
+camel.k.sources[2].type              = source
